+++ bindings/python/CHANGELOG.md
@@ -36,6 +36,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Unlock` types;
 - `PreparedTransactionData, SignedTransactionData, InputSigningData, RemainderData`;
 - `UtxoChanges`;
+- `TreasuryOutput, BasicOutput, AliasOutput, FoundryOutput, NftOutput`;
 
 ### Changed
 
@@ -62,6 +63,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Switched order of `AddressAndAmount` init params;
 - Renamed `PreparedTransactionData` to `PreparedTransaction`;
 - `{Client, SecretManager}::sign_transaction` return type from `SignedTransactionData` to `TransactionPayload`;
+- Split `Output` into multiple classes;
 
 ### Removed